+++ CHANGES.md
@@ -20,6 +20,7 @@ Changes in 6.9
 - Do not allow N/A for Period select in Courses.php
 - Deprecate UserCoursePeriodSchoolPeriod(), use UserPeriod() + UserCoursePeriod() instead in SchoolPeriodsSelectInput.fnc.php, StudentSummary.php, TakeAttendance.php & Current.php
 - Deprecate UserPeriod(), use per program logic instead in SchoolPeriodsSelectInput.fnc.php, StudentSummary.php, TakeAttendance.php & Current.php
+- HTML reorganize (Re)create Calendar form inputs in Calendar.php
 
 Changes in 6.8.1
 ----------------